backend/requirements.txt
"""
Requirements for TrueLine News Backend
"""

# Web Framework
Flask==2.3.2
Flask-CORS==4.0.0
flask-mongoengine==1.0.0

# Database
mongoengine==0.27.0
pymongo==4.4.0

# NLP and Text Processing
nltk==3.8.1
scikit-learn==1.3.0
beautifulsoup4==4.12.2
requests==2.31.0

# Environment and Configuration
python-dotenv==1.0.0

# Utilities
gunicorn==21.2.0

# Testing
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
